requirements.txt
# Framework Web
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Base de données
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentification et sécurité
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Validation
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# CORS
python-dotenv==1.0.0

# Utilitaires
python-dateutil==2.8.2